# market_go_api_gateway/api/docs/swagger.yaml
definitions:
...
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  sale_service.SaleProduct:
    properties:
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
      sale_id:
        type: string
    type: object
  sale_service.SaleProductCreateReq:
    properties:
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
      sale_id:
        type: string
    type: object
  sale_service.SaleProductCreateResp:
    properties:
      msg:
        type: string
    type: object
  sale_service.SaleProductDeleteResp:
    properties:
      msg:
        type: string
    type: object
  sale_service.SaleProductUpdateReq:
    properties:
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
      sale_id:
        type: string
    type: object
  sale_service.SaleProductUpdateResp:
    properties:
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/sale-products:
    get:
      consumes:
      - application/json
      description: Get sale-products
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        type: integer
      - default: 1
        description: page for response
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sale_service.SaleProduct'
            type: array
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: List sale-products
      tags:
      - Sale-products
    post:
      consumes:
      - application/json
      description: Create a new sale-product with the provided details
      parameters:
      - description: data of the sale-product
        in: body
        name: sale-product
        required: true
        schema:
          $ref: '#/definitions/sale_service.SaleProductCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sale_service.SaleProductCreateResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create a new sale-product
      tags:
      - Sale-products
  /v1/sale-products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a sale-product by its unique identifier
      parameters:
      - description: Sale ID to retrieve
        in: path
        name: sale_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sale_service.SaleProductDeleteResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete a sale-product
      tags:
      - Sale-products
    get:
      consumes:
      - application/json
      description: Retrieve a sale by its unique identifier
      parameters:
      - description: SaleProduct ID to retrieve
        in: path
        name: sale_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sale_service.SaleProduct'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get a sale_product by ID
      tags:
      - Sale-products
    put:
      consumes:
      - application/json
      description: Update an existing sale_product with the provided details
      parameters:
      - description: sale ID to update
        in: path
        name: sale_id
        required: true
        type: string
      - description: Updated data for the sale_product
        in: body
        name: sale_product
        required: true
        schema:
          $ref: '#/definitions/sale_service.SaleProductUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sale_service.SaleProductUpdateResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update an existing sale_product
      tags:
      - Sale-products
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"